+++ src/Framework/Entities/Components/Camera3DComponent.cpp
@@ -153,8 +153,8 @@ void Camera3DComponent::ComputeViewMatrix()
 void Camera3DComponent::ComputeProjectionMatrix()
 {
     m_needToComputeProjectionMatrix = false;
-	float ratio = static_cast<float>(m_Viewport.Width() * Game::Instance().GetWindowSize().x) / 
-		static_cast<float>(m_Viewport.Height() * Game::Instance().GetWindowSize().y);
+	float ratio = m_Viewport.Width() * Game::Instance().GetWindowSize().x / 
+		(m_Viewport.Height() * Game::Instance().GetWindowSize().y);
 
 	m_ProjectionMatrix.PerspectiveFOV(
 		m_fFOV,
